backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
celery>=5.3.0
redis>=4.6.0
anthropic>=0.5.0
python-dotenv>=1.0.0
sse-starlette>=1.6.0
typing-extensions>=4.7.0
pydantic>=2.10.0
pydantic-settings
asyncio>=3.4.3
cerebras_cloud_sdk>=1.26.0
google-genai>=1.7.0